Replaced int by size_t in all text and source files
git-svn-id: https://cppannotations.svn.sourceforge.net/svnroot/cppannotations/trunk@23 f6dd340e-d3f9-0310-b409-bdd246841980

+++ yo/classes/examples/getmembers.cc
@@ -16,7 +16,7 @@
        return d_phone;
     }
 
-    unsigned Person::weight() const
+    size_t Person::weight() const
     {
        return d_weight;
     }
