+++ yo/concrete/examples/fdout.h
@@ -9,7 +9,7 @@
 //=
     {
 //DATA
-        unsigned d_bufsize;
+        size_t d_bufsize;
         int     d_fd;
         char    *d_buffer;
 //=
@@ -21,7 +21,7 @@
                 d_bufsize(0),
                 d_buffer(0)
             {}
-            ofdnstreambuf(int fd, unsigned bufsize = 1)
+            ofdnstreambuf(int fd, size_t bufsize = 1)
             {
                 open(fd, bufsize);
             }
@@ -37,7 +37,7 @@
             }
 //=
 //OPEN
-            void open(int fd, unsigned bufsize = 1)
+            void open(int fd, size_t bufsize = 1)
             {
                 d_fd = fd;
                 d_bufsize = bufsize == 0 ? 1 : bufsize;
