+++ yo/concrete/fork2/ofdnstreambuf/open.cc
@@ -1,6 +1,6 @@
 #include "ofdnstreambuf.ih"
 
-void OFdNStreambuf::open(int fd, unsigned n)
+void OFdNStreambuf::open(int fd, size_t n)
 {
     d_fd = fd;
     d_n = (n == 0) ? 1 : n;
