+++ yo/concrete/fork2/pipe/readvia2.cc
@@ -1,10 +1,10 @@
 #include "pipe.ih"
 
-void Pipe::readVia(int const *fd, unsigned n)
+void Pipe::readVia(int const *fd, size_t n)
 {
     close(d_fd[WRITE]);
 
-    for (unsigned idx = 0; idx < n; idx++)
+    for (size_t idx = 0; idx < n; idx++)
         dup(READ, fd[idx]);
 
     close(d_fd[READ]);
