+++ yo/concrete/lexer/scanner/stacktrace.cc
@@ -2,7 +2,7 @@
 
     void Scanner::stackTrace()
     {
-        for (unsigned idx = 0; idx < d_fileName.size() - 1; ++idx)
+        for (size_t idx = 0; idx < d_fileName.size() - 1; ++idx)
             cerr << idx << ": " << d_fileName[idx]  << " included " <<
                                     d_fileName[idx + 1]  << endl;
     }
