+++ yo/concrete/streambuf/fdstreambuf/fdinmininum.cc
@@ -14,7 +14,7 @@ int main(int argc)
     string
         s;
 
-    for (unsigned line = 0; getline(is, s); line++)
+    for (size_t line = 0; getline(is, s); line++)
     {
         cout <<  s << endl;
         if ((line & 3) == 0)
